+++ packages/analytics-js-common/src/constants/QueueStatuses.ts
@@ -4,6 +4,7 @@ const QueueStatuses = {
   RECLAIM_START: 'reclaimStart',
   RECLAIM_END: 'reclaimEnd',
   ACK: 'ack',
+  BATCH_QUEUE: 'batchQueue',
 };
 
 export { QueueStatuses };