+++ packages/analytics-js-integrations/src/integrations/Amplitude/nativeSdkLoader.js
@@ -2,7 +2,7 @@ import { LOAD_ORIGIN } from '@rudderstack/analytics-js-common/v1.1/utils/constan
 // START-NO-SONAR-SCAN
 /* eslint-disable */
 
-function loader(e, t) {
+function loadNativeSdk(e, t) {
   var n = e.amplitude || { _q: [], _iq: {} };
   var r = t.createElement('script');
   r.type = 'text/javascript';
@@ -110,7 +110,5 @@ function loader(e, t) {
   e.amplitude = n;
 }
 window, document;
-/* eslint-enable */
-// END-NO-SONAR-SCAN
 
-export { loader };
+export { loadNativeSdk };