+++ packages/analytics-js-integrations/src/integrations/Klaviyo/index.js
@@ -1,3 +1,1 @@
-import Klaviyo from './browser';
-
-export { Klaviyo };
+export { default as Klaviyo } from './browser';