+++ packages/analytics-js-integrations/src/integrations/PinterestTag/index.js
@@ -1,3 +1,1 @@
-import PinterestTag from './browser';
-
-export { PinterestTag };
+export { default as PinterestTag } from './browser';