+++ packages/analytics-js-integrations/src/integrations/ProfitWell/index.js
@@ -1,3 +1,1 @@
-import ProfitWell from './browser';
-
-export { ProfitWell };
+export { default as ProfitWell } from './browser';