+++ packages/analytics-js-integrations/src/integrations/Qualaroo/utils.js
@@ -1,3 +1,4 @@
+/* eslint-disable func-names */
 /* eslint-disable no-underscore-dangle */
 import { getHashFromArray } from '../../utils/commonUtils';