+++ packages/analytics-js-integrations/src/integrations/SnapPixel/index.js
@@ -1,3 +1,1 @@
-import SnapPixel from './browser';
-
-export { SnapPixel };
+export { default as SnapPixel } from './browser';