+++ packages/analytics-js-integrations/src/integrations/Woopra/index.js
@@ -1,3 +1,1 @@
-import Woopra from './browser';
-
-export { Woopra };
+export { default as Woopra } from './browser';