+++ packages/analytics-js-plugins/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.0.0-beta.5](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.0.0-beta.4...@rudderstack/analytics-js-plugins@3.0.0-beta.5) (2023-08-21)
+
+**Note:** Version bump only for package @rudderstack/analytics-js-plugins
+
+
+
+
+
+# [3.0.0-beta.4](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.0.0-beta.3...@rudderstack/analytics-js-plugins@3.0.0-beta.4) (2023-08-17)
+
+**Note:** Version bump only for package @rudderstack/analytics-js-plugins
+
+
+
+
+
 # [3.0.0-beta.3](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-plugins@3.0.0-beta.2...@rudderstack/analytics-js-plugins@3.0.0-beta.3) (2023-08-10)